rename documents to documents_manager throughout the code
