+++ buster/examples/generate_embeddings.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from buster.documents import DeepLakeDocumentsManager
+from buster.documents_manager import DeepLakeDocumentsManager
 
 # Read the csv
 df = pd.read_csv("stackoverflow.csv")
